backend/ai_api/requirements.txt
flask
openai
python-dotenv
flask-cors
PyPDF2
SpeechRecognition
whisper
pydub
librosa
soundfile
numpy
torch
torchaudio
torchvision
